refactor(sensing): move the manual-paste heuristic into a sensing collector
