+++ extension/src/extension/services/telemetry/eventPipeline/compileEquivalentEmitter.ts
@@ -8,7 +8,6 @@ import {
     BuildResultClassification,
     CompileEquivalentEvent,
     DEFAULT_EQ_CONFIG,
-    DEFAULT_TRIGGER_CONFIG,
     EQConfig,
     ErrorSnapshot,
     SessionResettable,
@@ -178,13 +177,6 @@ export class CompileEquivalentEmitter implements vscode.Disposable, SessionReset
 }
 
 // ============================================================================
-// Constants
-// ============================================================================
-
-/** Character threshold above which a replaced range is likely formatter/refactoring, not paste */
-const FORMATTER_CHAR_THRESHOLD = 1000;
-
-// ============================================================================
 // Helper Functions
 // ============================================================================
 
@@ -236,32 +228,3 @@ function getErrorFamily(d: vscode.Diagnostic): string {
     return `${source}:${code}`;
 }
 
-/**
- * Detect if a text change is likely a manual paste (not formatter/copilot/snippet).
- * [Engineering heuristic — not paper-validated]
- *
- * From MVP Edge Case 3 (lines 739-758).
- */
-export function isLikelyManualPaste(
-    change: vscode.TextDocumentContentChangeEvent,
-    minLines: number = DEFAULT_TRIGGER_CONFIG.MULTILINE_PASTE_MIN_LINES,
-): boolean {
-    const insertedLines = change.text.split('\n').length;
-    if (insertedLines < minLines) {
-        return false;
-    }
-
-    // Formatter/refactoring: replaces large text range (>1000 chars)
-    if (change.rangeLength > FORMATTER_CHAR_THRESHOLD) {
-        return false;
-    }
-
-    // Copilot/snippet: replaces text on a single line with multi-line output
-    if (!change.range.isEmpty && change.range.isSingleLine) {
-        return false;
-    }
-
-    // Pure insert (range.isEmpty) → likely paste (Ctrl+V)
-    // Multi-line replacement (range spans multiple lines, ≤1000 chars) → likely paste-over-selection
-    return true;
-}
