+++ extension/test/logic/telemetry/manualPaste.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest';
 
-import { isLikelyManualPaste } from '@extension/services/telemetry/eventPipeline/compileEquivalentEmitter';
+import { isLikelyManualPaste } from '@extension/services/sensing/collectors/paste';
 
 type Change = Parameters<typeof isLikelyManualPaste>[0];
