+++ extension/test/unit/struggle-detection/boundaryTriggerAndCadence.test.ts
@@ -13,8 +13,8 @@
 import * as assert from 'assert';
 import * as sinon from 'sinon';
 
+import { isLikelyManualPaste } from '@extension/services/sensing/collectors/paste';
 import { BoundaryTriggerEmitter } from '@extension/services/telemetry/eventPipeline/boundaryTriggerEmitter';
-import { isLikelyManualPaste } from '@extension/services/telemetry/eventPipeline/compileEquivalentEmitter';
 import { InactivityService } from '@extension/services/telemetry/inactivityService';
 import { AdaptiveCadence } from '@extension/services/telemetry/intervention/adaptiveCadence';
 import { DEFAULT_TRIGGER_CONFIG, TriggerType } from '@extension/services/telemetry/types';
